Add: onboarding ui, personalize ui, network layer, dependency injection,  some resources
