--- app/src/main/java/com/sci/recipeandroid/feature/onboarding/domain/model/AllergiesIngredientModel.kt
@@ -1,6 +1,0 @@
-package com.sci.recipeandroid.feature.onboarding.domain.model
-
-data class AllergiesIngredientModel(
-    val id:String,
-    val name:String?
-)
