--- app/src/main/java/com/sci/recipeandroid/feature/onboarding/domain/model/DietRecipeModel.kt
@@ -1,7 +1,0 @@
-package com.sci.recipeandroid.feature.onboarding.domain.model
-
-data class DietRecipeModel(
-    val id:String,
-    val imageUrl:String?,
-    val name:String?
-)
