--- app/src/main/java/com/sci/recipeandroid/feature/onboarding/domain/model/PersonalizeDataModel.kt
@@ -1,9 +1,0 @@
-package com.sci.recipeandroid.feature.onboarding.domain.model
-
-import com.sci.recipeandroid.feature.onboarding.data.model.response.AllergiesIngredient
-import com.sci.recipeandroid.feature.onboarding.data.model.response.DietRecipe
-
-data class PersonalizeDataModel (
-    val dietRecipe:List<DietRecipeModel>,
-    val allergiesIngredient:List<AllergiesIngredientModel>
-)
