--- app/src/main/java/com/sci/recipeandroid/feature/onboarding/ui/fragment/PersonalizeGoalsFragment.kt
@@ -1,26 +1,0 @@
-package com.sci.recipeandroid.feature.onboarding.ui.fragment
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.sci.recipeandroid.databinding.FragmentPersonalizeGoalsBinding
-
-class PersonalizeGoalsFragment : Fragment() {
-    private lateinit var binding: FragmentPersonalizeGoalsBinding
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentPersonalizeGoalsBinding.inflate(inflater, container, false)
-        // Inflate the layout for this fragment
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-    }
-}
