--- app/src/main/java/com/sci/recipeandroid/feature/onboarding/ui/viewholder/OnBoardViewHolder.kt
@@ -1,12 +1,0 @@
-package com.sci.recipeandroid.feature.onboarding.ui.viewholder
-
-import androidx.recyclerview.widget.RecyclerView
-import com.sci.recipeandroid.databinding.ItemViewOnboardBinding
-import com.sci.recipeandroid.feature.onboarding.ui.model.OnBoardModel
-
-class OnBoardViewHolder(private val binding: ItemViewOnboardBinding) : RecyclerView.ViewHolder(binding.root) {
-
-    fun bind(onBoardModel: OnBoardModel) {
-
-    }
-}
