+++ app/src/main/java/com/sci/recipeandroid/feature/personalize/data/model/response/PersonalizeResponse.kt
@@ -1,0 +1,30 @@
+package com.sci.recipeandroid.feature.personalize.data.model.response
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PersonalizeResponse(
+    val code:Int?,
+    val message:String?,
+    val data: PersonalizeData
+)
+
+
+@Serializable
+data class PersonalizeData(
+    val dietRecipe:List<DietRecipe>,
+    val allergiesIngredient:List<AllergiesIngredient>
+)
+
+@Serializable
+data class DietRecipe (
+    val id:Int,
+    val imageUrl:String?,
+    val name:String?
+)
+
+@Serializable
+data class AllergiesIngredient (
+    val id:String,
+    val name:String?
+)
