+++ app/src/main/java/com/sci/recipeandroid/feature/personalize/di/NetworkModule.kt
@@ -1,0 +1,21 @@
+package com.sci.recipeandroid.feature.personalize.di
+
+import com.sci.recipeandroid.feature.personalize.domain.service.PersonalizeService
+import com.sci.recipeandroid.feature.personalize.network.KtorHelper
+import org.koin.dsl.module
+
+val networkModule = module {
+    single {
+        KtorHelper().getClient()
+    }
+
+    single {
+        PersonalizeService(
+            get()
+        )
+    }
+}
+
+
+
+
