+++ app/src/main/java/com/sci/recipeandroid/feature/personalize/di/ViewModelModule.kt
@@ -1,0 +1,14 @@
+package com.sci.recipeandroid.feature.personalize.di
+
+import com.sci.recipeandroid.feature.personalize.ui.viewmodel.PersonalizeViewModel
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.dsl.module
+
+val viewModelModule = module {
+    viewModel {
+        PersonalizeViewModel(
+            get()
+        )
+    }
+}
+
