+++ app/src/main/java/com/sci/recipeandroid/feature/personalize/domain/model/AllergiesIngredientModel.kt
@@ -1,0 +1,6 @@
+package com.sci.recipeandroid.feature.personalize.domain.model
+
+data class AllergiesIngredientModel(
+    val id:String,
+    val name:String?
+)
