+++ app/src/main/java/com/sci/recipeandroid/feature/personalize/domain/model/DietRecipeModel.kt
@@ -1,0 +1,7 @@
+package com.sci.recipeandroid.feature.personalize.domain.model
+
+data class DietRecipeModel(
+    val id:Int,
+    val imageUrl:String?,
+    val name:String?
+)
