+++ app/src/main/java/com/sci/recipeandroid/feature/personalize/domain/model/PersonalizeDataModel.kt
@@ -1,0 +1,6 @@
+package com.sci.recipeandroid.feature.personalize.domain.model
+
+data class PersonalizeDataModel (
+    val dietRecipe:List<DietRecipeModel>,
+    val allergiesIngredient:List<AllergiesIngredientModel>
+)
