+++ app/src/main/java/com/sci/recipeandroid/feature/personalize/network/KtorHelper.kt
@@ -1,0 +1,24 @@
+package com.sci.recipeandroid.feature.personalize.network
+
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.android.Android
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.plugins.defaultRequest
+import io.ktor.serialization.kotlinx.json.json
+import kotlinx.serialization.json.Json
+
+class KtorHelper {
+    fun getClient(): HttpClient {
+        return HttpClient(Android) {
+            defaultRequest {
+               url()
+            }
+            install(ContentNegotiation) {
+                json(Json {
+                    ignoreUnknownKeys = true
+                })
+            }
+        }
+    }
+}
+/**/
