+++ app/src/main/java/com/sci/recipeandroid/feature/personalize/ui/viewholder/DietRecipeViewHolder.kt
@@ -1,0 +1,32 @@
+package com.sci.recipeandroid.feature.personalize.ui.viewholder
+
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
+import com.sci.recipeandroid.R
+import com.sci.recipeandroid.databinding.ItemViewDietRecipeBinding
+import com.sci.recipeandroid.feature.personalize.domain.model.DietRecipeModel
+
+class DietRecipeViewHolder(private val binding: ItemViewDietRecipeBinding,private val onClickItem: (Int) -> Unit) : RecyclerView.ViewHolder(binding.root) {
+
+    fun bind(model: DietRecipeModel, isSelected: Boolean) {
+        binding.apply {
+            tvRecipeDiet.text = model.name
+            Glide.with(itemView.context)
+                .load(model.imageUrl)
+        }
+
+        if (isSelected) {
+            binding.cardDietRecipe.strokeColor = ContextCompat.getColor(itemView.context, R.color.color_primary)
+        } else {
+            binding.cardDietRecipe.strokeColor = ContextCompat.getColor(itemView.context, R.color.color_surface)
+        }
+
+        itemView.setOnClickListener {
+
+            onClickItem(
+                adapterPosition
+            )
+        }
+    }
+}
