+++ app/src/main/java/com/sci/recipeandroid/feature/personalize/ui/viewholder/OnBoardViewHolder.kt
@@ -1,0 +1,12 @@
+package com.sci.recipeandroid.feature.personalize.ui.viewholder
+
+import androidx.recyclerview.widget.RecyclerView
+import com.sci.recipeandroid.databinding.ItemViewOnboardBinding
+import com.sci.recipeandroid.feature.onboarding.ui.model.OnBoardModel
+
+class OnBoardViewHolder(private val binding: ItemViewOnboardBinding) : RecyclerView.ViewHolder(binding.root) {
+
+    fun bind(onBoardModel: OnBoardModel) {
+
+    }
+}
