+++ app/src/main/res/layout/fragment_onboard.xml
@@ -8,8 +8,8 @@
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/ivOnBoardImage"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
+        android:layout_width="@dimen/margin_none"
+        android:layout_height="@dimen/margin_none"
         android:scaleType="centerCrop"
         app:layout_constraintDimensionRatio="W,2:1.7"
         app:layout_constraintEnd_toEndOf="parent"
@@ -18,7 +18,7 @@
         tools:srcCompat="@drawable/onboard2" />
 
     <ImageView
-        android:id="@+id/tvlogo"
+        android:id="@+id/ivlogo"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:src="@drawable/app_logo_square"
@@ -46,7 +46,7 @@
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/tvSubtitle"
         style="@style/CustomFont_Black_Regular"
-        android:layout_width="0dp"
+        android:layout_width="@dimen/margin_none"
         android:layout_height="wrap_content"
         tools:text="@string/onboarding1_text"
         android:textAlignment="center"
@@ -60,7 +60,7 @@
         style="@style/CustomFont_Login_Regular"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        tools:text="@string/lbl_login"
+        android:text="@string/lbl_login"
         android:textStyle="bold"
         android:textAllCaps="true"
         android:textAlignment="center"
@@ -74,7 +74,7 @@
     <LinearLayout
         android:id="@+id/slider_dots"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
+        android:layout_height="@dimen/margin_none"
         android:gravity="center_vertical|center_horizontal"
         android:orientation="horizontal"
         app:layout_constraintTop_toTopOf="@id/btnNext"
