+++ app/src/main/res/layout/fragment_personalize_goals.xml
@@ -3,6 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="@dimen/margin_medium_2"
+    android:layout_margin="@dimen/margin_medium_2"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:fitsSystemWindows="true">
