+++ app/src/main/res/layout/item_view_diet_recipe.xml
@@ -1,23 +1,41 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="@dimen/recipe_diet_width"
+
+<com.google.android.material.card.MaterialCardView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/card_diet_recipe"
+    android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:background="@drawable/ic_recipe_diet_bg"
-    android:id="@+id/container">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    app:strokeColor="@color/color_surface"
+    app:cardElevation="@dimen/margin_small"
+    app:cardCornerRadius="@dimen/margin_medium_x_12"
+    app:cardUseCompatPadding="true"
+    app:cardPreventCornerOverlap="true"
+    app:cardBackgroundColor="@color/color_surface">
 
-    <ImageView
-        android:id="@+id/ivRecipeDiet"
-        android:layout_width="@dimen/recipe_diet_width_height"
-        android:layout_height="@dimen/recipe_diet_width_height"
-        android:src="@drawable/none"
-        android:layout_margin="5dp"/>
-
-    <TextView
-        android:id="@+id/tvRecipeDiet"
+    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/lbl_none"
-        android:layout_gravity="center"/>
+        android:padding="@dimen/margin_card_medium"
+        android:orientation="vertical"
+        android:id="@+id/container">
 
-</LinearLayout>
+        <ImageView
+            android:id="@+id/ivRecipeDiet"
+            android:layout_width="@dimen/recipe_diet_width_height"
+            android:layout_height="@dimen/recipe_diet_width_height"
+            android:scaleType="fitXY"
+            android:src="@drawable/none"/>
+
+        <TextView
+            android:id="@+id/tvRecipeDiet"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/lbl_none"
+            android:layout_gravity="center"
+            android:paddingTop="@dimen/margin_medium"
+            style="@style/CustomFont_Small_Semi_Bold"/>
+
+    </LinearLayout>
+
+</com.google.android.material.card.MaterialCardView>
+
