+++ app/src/main/res/navigation/navigation_graph.xml
@@ -18,6 +18,41 @@
         android:id="@+id/onBoardFragment"
         android:name="com.sci.recipeandroid.feature.onboarding.ui.fragment.OnBoardFragment"
         android:label="OnBoardFragment"
-        tools:layout="@layout/item_view_onboard"/>
+        tools:layout="@layout/item_view_onboard">
+
+        <argument
+            android:name="onboardStep"
+            app:argType="integer"
+            android:defaultValue="1" />
+        <action
+            android:id="@+id/action_onboardFragment_to_nextOnboardFragment"
+            app:destination="@id/onBoardFragment">
+            <argument
+                android:name="onboardStep"
+                app:argType="integer"
+                android:defaultValue="2" />
+        </action>
+        <action
+            android:id="@+id/action_onBoardFragment_to_personalizeOptionsFragment"
+            app:destination="@id/personalizeOptionsFragment" />
+
+    </fragment>
+    <fragment
+        android:id="@+id/personalizeOptionsFragment"
+        android:name="com.sci.recipeandroid.feature.personalize.ui.fragment.PersonalizeOptionsFragment"
+        android:label="fragment_personalize_options"
+        tools:layout="@layout/fragment_personalize_options" >
+        <action
+            android:id="@+id/action_personalizeOptionsFragment_to_personalizeGoalsFragment"
+            app:destination="@id/personalizeGoalsFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/personalizeGoalsFragment"
+        android:name="com.sci.recipeandroid.feature.personalize.ui.fragment.PersonalizeGoalsFragment"
+        android:label="PersonalizeGoalsFragment" >
+        <action
+            android:id="@+id/action_personalizeGoalsFragment_to_personalizeOptionsFragment"
+            app:destination="@id/personalizeOptionsFragment" />
+    </fragment>
 
 </navigation>
