+++ app/src/main/res/values/dimens.xml
@@ -35,7 +35,10 @@
     <dimen name="margin_large_n">-24dp</dimen>
     <dimen name="margin_xlarge_n">-32dp</dimen>
 
+    <dimen name="value_padding">10dp</dimen>
+
     <dimen name="text_smaller">10sp</dimen>
+    <dimen name="text_smaller_2">11sp</dimen>
     <dimen name="text_small">12sp</dimen>
     <dimen name="text_regular">14sp</dimen>
     <dimen name="text_regular_2x">16sp</dimen>
