+++ app/src/main/res/values/themes.xml
@@ -4,10 +4,6 @@
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
 
-        <item name="android:windowTranslucentStatus">true</item>
-        <item name="android:windowTranslucentNavigation">true</item>
-        <item name="windowActionBar">false</item>
-        <item name="windowNoTitle">true</item>
     </style>
 
     <style name="Theme.RecipeAndroid" parent="Base.Theme.RecipeAndroid" />
@@ -38,5 +34,18 @@
         <item name="fontFamily">@font/poppins_semi_bold</item>
         <item name="android:textSize">@dimen/text_regular</item>
         <item name="android:textColor">@color/color_button_label</item>
+    </style>
+
+    <style name="CustomFont_Small_Semi_Bold">
+        <item name="android:fontFamily">@font/poppins_semi_bold</item>
+        <item name="fontFamily">@font/poppins_semi_bold</item>
+        <item name="android:textSize">@dimen/text_smaller_2</item>
+        <item name="android:textColor">@color/color_text_primary</item>
+    </style>
+
+    <style name="CustomFont_Tag_Small_Semi_Bold">
+        <item name="android:fontFamily">@font/poppins_semi_bold</item>
+        <item name="fontFamily">@font/poppins_semi_bold</item>
+        <item name="android:textSize">@dimen/text_regular</item>
     </style>
 </resources>
